nextLink should be used directly
I was unable to find verification of this pattern for AAD Graph, but
with MS Graph, nextLink should be treated as an opaque URL. You should
never parse out the skip token and attempt to create the paging URL
yourself. If this is true for AAD Graph as well, this will need to be
updated in several places.